@@ -1,24 +1,20 @@
 package org.jenkinsci.plugins.ghprb;
 
-import java.io.IOException;
-
 import hudson.Extension;
-import hudson.Launcher;
-import hudson.model.AbstractBuild;
-import hudson.model.BuildListener;
-import hudson.model.Environment;
 import hudson.model.Run;
 import hudson.model.TaskListener;
 import hudson.model.listeners.RunListener;
+
+import javax.annotation.Nonnull;
 
 /**
  * @author janinko
  */
 @Extension
-public class GhprbBuildListener extends RunListener<AbstractBuild<?, ?>> {
+public class GhprbBuildListener extends RunListener<Run<?, ?>> {
 
     @Override
-    public void onStarted(AbstractBuild<?, ?> build, TaskListener listener) {
+    public void onStarted(Run<?, ?> build, TaskListener listener) {
         GhprbTrigger trigger = Ghprb.extractTrigger(build);
         if (trigger != null && trigger.getBuilds() != null) {
             trigger.getBuilds().onStarted(build, listener);
@@ -26,20 +22,10 @@ public class GhprbBuildListener extends RunListener<AbstractBuild<?, ?>> {
     }
 
     @Override
-    public void onCompleted(AbstractBuild<?, ?> build, TaskListener listener) {
+    public void onCompleted(Run<?, ?> build, @Nonnull TaskListener listener) {
         GhprbTrigger trigger = Ghprb.extractTrigger(build);
         if (trigger != null && trigger.getBuilds() != null) {
             trigger.getBuilds().onCompleted(build, listener);
         }
-    }
-
-    @Override
-    public Environment setUpEnvironment(@SuppressWarnings("rawtypes") AbstractBuild build, Launcher launcher, BuildListener listener) throws IOException, InterruptedException, Run.RunnerAbortedException {
-        GhprbTrigger trigger = Ghprb.extractTrigger(build);
-        if (trigger != null && trigger.getBuilds() != null) {
-            trigger.getBuilds().onEnvironmentSetup(build, launcher, listener);
-        }
-
-        return new hudson.model.Environment(){};
     }
 }
