@@ -1,10 +1,9 @@
 package org.jenkinsci.plugins.ghprb;
 
+import hudson.model.Cause;
 import org.apache.commons.lang.StringUtils;
 import org.kohsuke.github.GHUser;
 import org.kohsuke.github.GitUser;
-
-import hudson.model.Cause;
 
 import java.net.URL;
 
@@ -12,39 +11,57 @@ import java.net.URL;
  * @author Honza Brázdil
  */
 public class GhprbCause extends Cause {
+
+    public static final int MAX_WIDTH = 30;
+
     private final String commit;
+
     private final int pullID;
+
     private final boolean merged;
+
     private final String targetBranch;
+
     private final String sourceBranch;
+
     private final String authorEmail;
+
     private final String title;
+
     private final URL url;
-    private final GHUser triggerSender;
+
+    private final transient GHUser triggerSender;
+
     private final String commentBody;
-    private final GitUser commitAuthor;
-    private final GHUser pullRequestAuthor;
+
+    private final transient GitUser commitAuthor;
+
+    private final transient GHUser pullRequestAuthor;
+
     private final String description;
+
     private final String authorRepoGitUrl;
+
     private final String repoName;
+
     private final String credentialsId;
 
-    public GhprbCause(String commit, 
-            int pullID, 
-            boolean merged, 
-            String targetBranch, 
-            String sourceBranch, 
-            String authorEmail, 
-            String title, 
-            URL url, 
-            GHUser triggerSender, 
-            String commentBody,
-            GitUser commitAuthor,
-            GHUser pullRequestAuthor,
-            String description,
-            String authorRepoGitUrl,
-            String repoName,
-            String credentialsId) {
+    public GhprbCause(String commit,
+                      int pullID,
+                      boolean merged,
+                      String targetBranch,
+                      String sourceBranch,
+                      String authorEmail,
+                      String title,
+                      URL url,
+                      GHUser triggerSender,
+                      String commentBody,
+                      GitUser commitAuthor,
+                      GHUser pullRequestAuthor,
+                      String description,
+                      String authorRepoGitUrl,
+                      String repoName,
+                      String credentialsId) {
 
         this.commit = commit;
         this.pullID = pullID;
@@ -64,7 +81,7 @@ public class GhprbCause extends Cause {
         this.repoName = repoName;
         this.credentialsId = credentialsId;
     }
-    
+
     @Override
     public String getShortDescription() {
         return "GitHub pull request #" + pullID + " of commit " + commit + (merged ? ", no merge conflicts." : ", has merge conflicts.");
@@ -111,7 +128,6 @@ public class GhprbCause extends Cause {
     }
 
     /**
-     * 
      * @return the title of the cause, not null.
      */
     public String getTitle() {
@@ -119,18 +135,16 @@ public class GhprbCause extends Cause {
     }
 
     /**
-     * Returns
-     * 
-     * @return  at most the first 30 characters of the title
+     * @return at most the first 30 characters of the title
      */
     public String getAbbreviatedTitle() {
-        return StringUtils.abbreviate(getTitle(), 30);
+        return StringUtils.abbreviate(getTitle(), MAX_WIDTH);
     }
 
     public GitUser getCommitAuthor() {
         return commitAuthor;
     }
-    
+
     public String getDescription() {
         return description;
     }
