@@ -1,19 +1,20 @@
 package org.jenkinsci.plugins.ghprb;
 
+import org.kohsuke.github.GHOrganization;
+import org.kohsuke.github.GHUser;
+
 import java.io.IOException;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-
-import org.kohsuke.github.GHOrganization;
-import org.kohsuke.github.GHUser;
 
 /**
  * @author janinko
  */
 public class GhprbGitHub {
-    private static final Logger logger = Logger.getLogger(GhprbGitHub.class.getName());
+    private static final Logger LOGGER = Logger.getLogger(GhprbGitHub.class.getName());
+
     private final GhprbTrigger trigger;
-    
+
     public GhprbGitHub(GhprbTrigger trigger) {
         this.trigger = trigger;
     }
@@ -23,11 +24,11 @@ public class GhprbGitHub {
         try {
             GHOrganization org = trigger.getGitHub().getOrganization(organisation);
             orgHasMember = org.hasMember(member);
-            logger.log(Level.FINE, "org.hasMember(member)? user:{0} org: {1} == {2}", 
-                    new Object[] { member.getLogin(), organisation, orgHasMember ? "yes" : "no" });
+            LOGGER.log(Level.FINE, "org.hasMember(member)? user:{0} org: {1} == {2}",
+                    new Object[] {member.getLogin(), organisation, orgHasMember ? "yes" : "no"});
 
         } catch (IOException ex) {
-            logger.log(Level.SEVERE, null, ex);
+            LOGGER.log(Level.SEVERE, null, ex);
             return false;
         }
         return orgHasMember;
@@ -37,7 +38,7 @@ public class GhprbGitHub {
         try {
             return trigger.getGitHub().getMyself().getLogin();
         } catch (IOException ex) {
-            logger.log(Level.SEVERE, null, ex);
+            LOGGER.log(Level.SEVERE, null, ex);
             return null;
         }
     }
