@@ -1,11 +1,10 @@
 package org.jenkinsci.plugins.ghprb;
 
 import hudson.Extension;
-import hudson.model.AbstractProject;
+import hudson.model.Job;
 import hudson.model.UnprotectedRootAction;
 import hudson.security.ACL;
 import hudson.security.csrf.CrumbExclusion;
-
 import org.acegisecurity.Authentication;
 import org.acegisecurity.context.SecurityContextHolder;
 import org.apache.commons.io.IOUtils;
@@ -17,6 +16,10 @@ import org.kohsuke.github.GitHub;
 import org.kohsuke.stapler.StaplerRequest;
 import org.kohsuke.stapler.StaplerResponse;
 
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.StringReader;
@@ -31,20 +34,20 @@ import java.util.concurrent.Executors;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpServletRequest;
-
 /**
  * @author Honza Brázdil jbrazdil@redhat.com
  */
 @Extension
 public class GhprbRootAction implements UnprotectedRootAction {
+
     static final String URL = "ghprbhook";
-    private static final Logger logger = Logger.getLogger(GhprbRootAction.class.getName());
-    
+
+    private static final Logger LOGGER = Logger.getLogger(GhprbRootAction.class.getName());
+
+    private static final int PAYLOAD_LENGTH = 8;
+
     private Set<StartTrigger> triggerThreads;
+
     private ExecutorService pool;
 
     public String getIconFileName() {
@@ -58,11 +61,11 @@ public class GhprbRootAction implements UnprotectedRootAction {
     public String getUrlName() {
         return URL;
     }
-    
+
     public int getThreadCount() {
         return triggerThreads == null ? 0 : triggerThreads.size();
     }
-    
+
     public GhprbRootAction() {
         triggerThreads = Collections.newSetFromMap(new WeakHashMap<StartTrigger, Boolean>());
         this.pool = Executors.newCachedThreadPool();
@@ -79,40 +82,40 @@ public class GhprbRootAction implements UnprotectedRootAction {
         if (type != null && type.toLowerCase().startsWith("application/json")) {
             body = extractRequestBody(req);
             if (body == null) {
-                logger.log(Level.SEVERE, "Can't get request body for application/json.");
+                LOGGER.log(Level.SEVERE, "Can't get request body for application/json.");
                 resp.setStatus(StaplerResponse.SC_BAD_REQUEST);
                 return;
             }
             payload = body;
         } else if (type != null && type.toLowerCase().startsWith("application/x-www-form-urlencoded")) {
             body = extractRequestBody(req);
-            if (body == null || body.length() <= 8) {
-                logger.log(Level.SEVERE,
-                           "Request doesn't contain payload. "
-                                         + "You're sending url encoded request, so you should pass github payload through 'payload' request parameter");
+            if (body == null || body.length() <= PAYLOAD_LENGTH) {
+                LOGGER.log(Level.SEVERE,
+                        "Request doesn't contain payload. You are sending url encoded request, "
+                        + "so you should pass github payload through 'payload' request parameter");
                 resp.setStatus(StaplerResponse.SC_BAD_REQUEST);
                 return;
             }
             try {
                 String encoding = req.getCharacterEncoding();
-                payload = URLDecoder.decode(body.substring(8), encoding != null ? encoding : "UTF-8");
+                payload = URLDecoder.decode(body.substring(PAYLOAD_LENGTH), encoding != null ? encoding : "UTF-8");
             } catch (UnsupportedEncodingException e) {
-                logger.log(Level.SEVERE, "Error while trying to decode the payload");
+                LOGGER.log(Level.SEVERE, "Error while trying to decode the payload");
                 resp.setStatus(StaplerResponse.SC_BAD_REQUEST);
                 return;
             }
         }
 
         if (payload == null) {
-            logger.log(Level.SEVERE,
-                       "Payload is null, maybe content type ''{0}'' is not supported by this plugin. "
-                                     + "Please use 'application/json' or 'application/x-www-form-urlencoded'",
-                       new Object[] { type });
+            LOGGER.log(Level.SEVERE,
+                    "Payload is null, maybe content type ''{0}'' is not supported by this plugin. "
+                            + "Please use 'application/json' or 'application/x-www-form-urlencoded'",
+                    new Object[] {type});
             resp.setStatus(StaplerResponse.SC_UNSUPPORTED_MEDIA_TYPE);
             return;
         }
 
-        logger.log(Level.FINE, "Got payload event: {0}", event);
+        LOGGER.log(Level.FINE, "Got payload event: {0}", event);
         final String threadBody = body;
         final String threadPayload = payload;
         handleAction(event, signature, threadPayload, threadBody);
@@ -140,30 +143,30 @@ public class GhprbRootAction implements UnprotectedRootAction {
                 GHIssueState state = comment.getIssue().getState();
 
                 if (state == GHIssueState.CLOSED) {
-                    logger.log(Level.INFO, "Skip comment on closed PR");
+                    LOGGER.log(Level.INFO, "Skip comment on closed PR");
                     return;
                 }
 
                 if (!comment.getIssue().isPullRequest()) {
-                    logger.log(Level.INFO, "Skip comment on Issue");
+                    LOGGER.log(Level.INFO, "Skip comment on Issue");
                     return;
                 }
 
                 repoName = comment.getRepository().getFullName();
 
-                logger.log(Level.INFO,
-                           "Checking issue comment ''{0}'' for repo {1}",
-                           new Object[] { comment.getComment().getBody(), repoName });
+                LOGGER.log(Level.INFO,
+                        "Checking issue comment ''{0}'' for repo {1}",
+                        new Object[] {comment.getComment().getBody(), repoName});
 
             } else if (StringUtils.equalsIgnoreCase("pull_request", event)) {
 
                 pr = getPullRequest(payload, gh);
                 repoName = pr.getRepository().getFullName();
 
-                logger.log(Level.INFO, "Checking PR #{1} for {0}", new Object[] { repoName, pr.getNumber() });
+                LOGGER.log(Level.INFO, "Checking PR #{1} for {0}", new Object[] {repoName, pr.getNumber()});
 
             } else {
-                logger.log(Level.WARNING, "Request not known for event: {0}", new Object[] { event });
+                LOGGER.log(Level.WARNING, "Request not known for event: {0}", new Object[] {event});
                 return;
             }
 
@@ -172,16 +175,18 @@ public class GhprbRootAction implements UnprotectedRootAction {
             handleEvent(triggers, payload, pr, comment);
 
         } catch (IOException e) {
-            logger.log(Level.SEVERE, "Unable to connect to GitHub anonymously", e);
+            LOGGER.log(Level.SEVERE, "Unable to connect to GitHub anonymously", e);
         } finally {
             SecurityContextHolder.getContext().setAuthentication(old);
         }
     }
 
     private class StartTrigger implements Runnable {
-        GhprbTrigger trigger;
-        PullRequest pr;
-        IssueComment comment;
+        private GhprbTrigger trigger;
+
+        private PullRequest pr;
+
+        private IssueComment comment;
 
         @Override
         public void run() {
@@ -194,7 +199,7 @@ public class GhprbRootAction implements UnprotectedRootAction {
                     triggerComment(trigger, comment);
                 }
             } catch (Exception e) {
-                logger.log(Level.SEVERE, "Failed to run thread", e);
+                LOGGER.log(Level.SEVERE, "Failed to run thread", e);
             } finally {
                 triggerThreads.remove(this);
             }
@@ -220,7 +225,7 @@ public class GhprbRootAction implements UnprotectedRootAction {
                 triggerThreads.add(runner);
                 pool.submit(runner);
             } catch (IOException e) {
-                logger.log(Level.SEVERE, "Unable to get authorized version of event", e);
+                LOGGER.log(Level.SEVERE, "Unable to get authorized version of event", e);
             }
         }
     }
@@ -238,7 +243,7 @@ public class GhprbRootAction implements UnprotectedRootAction {
                     sb.append(", repo: ");
                     sb.append(comment.getRepository().getFullName());
 
-                    logger.log(Level.SEVERE, sb.toString(), e);
+                    LOGGER.log(Level.SEVERE, sb.toString(), e);
                 }
             }
         }.start();
@@ -250,13 +255,13 @@ public class GhprbRootAction implements UnprotectedRootAction {
             public void run() {
                 try {
                     trigger.handlePR(pr);
-                } catch (Exception e) {
+                } catch (Throwable th) {
                     StringBuilder sb = new StringBuilder();
                     sb.append("Unable to handle PR# ");
                     sb.append(pr.getNumber());
                     sb.append(" for repo: ");
                     sb.append(pr.getRepository().getFullName());
-                    logger.log(Level.SEVERE, sb.toString(), e);
+                    LOGGER.log(Level.SEVERE, sb.toString(), th);
                 }
             }
         }.start();
@@ -270,8 +275,7 @@ public class GhprbRootAction implements UnprotectedRootAction {
 
     private IssueComment getIssueComment(String payload,
                                          GitHub gh) throws IOException {
-        IssueComment issueComment = gh.parseEventPayload(new StringReader(payload), IssueComment.class);
-        return issueComment;
+        return gh.parseEventPayload(new StringReader(payload), IssueComment.class);
     }
 
     private String extractRequestBody(StaplerRequest req) {
@@ -293,13 +297,13 @@ public class GhprbRootAction implements UnprotectedRootAction {
                                           String signature) {
         Set<GhprbTrigger> triggers = new HashSet<GhprbTrigger>();
 
-        Set<AbstractProject<?, ?>> projects = GhprbTrigger.getDscp().getRepoTriggers(repoName);
+        Set<Job<?, ?>> projects = GhprbTrigger.getDscp().getRepoTriggers(repoName);
         if (projects != null) {
-            for (AbstractProject<?, ?> project : projects) {
+            for (Job<?, ?> project : projects) {
                 GhprbTrigger trigger = Ghprb.extractTrigger(project);
                 if (trigger == null) {
-                    logger.log(Level.WARNING,
-                               "Warning, trigger unexpectedly null for project " + project.getFullName());
+                    LOGGER.log(Level.WARNING,
+                            "Warning, trigger unexpectedly null for project " + project.getFullName());
                     continue;
                 }
                 try {
@@ -307,9 +311,9 @@ public class GhprbRootAction implements UnprotectedRootAction {
                         triggers.add(trigger);
                     }
                 } catch (Exception e) {
-                    logger.log(Level.SEVERE,
-                               "Failed to match signature for trigger on project: " + trigger.getProjectName(),
-                               e);
+                    LOGGER.log(Level.SEVERE,
+                            "Failed to match signature for trigger on project: " + trigger.getProjectName(),
+                            e);
                 }
             }
         }
