@@ -1,13 +1,14 @@
 package org.jenkinsci.plugins.ghprb;
 
+import com.google.common.collect.ListMultimap;
 import hudson.Extension;
-import hudson.model.TaskListener;
 import hudson.model.AbstractBuild;
-
+import hudson.model.TaskListener;
 import org.jenkinsci.plugins.tokenmacro.DataBoundTokenMacro;
 import org.jenkinsci.plugins.tokenmacro.MacroEvaluationException;
 
 import java.io.IOException;
+import java.util.Map;
 
 /**
  * {@code PR_Name} token that expands to the PR Name. {@code PR_User} token that expands to the PR Opener's email.
@@ -22,8 +23,22 @@ public class GhprbTokenMacro extends DataBoundTokenMacro {
     }
 
     @Override
-    public String evaluate(AbstractBuild<?, ?> context, TaskListener listener, String macroName) 
-            throws MacroEvaluationException, IOException, InterruptedException {
+    public String evaluate(
+            AbstractBuild<?, ?> abstractBuild,
+            TaskListener taskListener,
+            String s
+    ) throws MacroEvaluationException, IOException, InterruptedException {
+        return null;
+    }
+
+    @Override
+    public String evaluate(
+            AbstractBuild<?, ?> context,
+            TaskListener listener,
+            String macroName,
+            Map<String, String> arguments,
+            ListMultimap<String, String> argumentMultimap
+    ) throws MacroEvaluationException, IOException, InterruptedException {
         GhprbCause cause = (GhprbCause) context.getCause(GhprbCause.class);
         if (cause == null) {
             return "";
