@@ -1,9 +1,10 @@
+// CHECKSTYLE:OFF
 package org.jenkinsci.plugins.ghprb;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-
+import antlr.ANTLRException;
+import hudson.model.Job;
+import hudson.triggers.Trigger;
+import hudson.util.DescribableList;
 import org.apache.commons.lang.StringUtils;
 import org.jenkinsci.plugins.ghprb.extensions.GhprbExtension;
 import org.jenkinsci.plugins.ghprb.extensions.GhprbExtensionDescriptor;
@@ -13,58 +14,60 @@ import org.jenkinsci.plugins.ghprb.extensions.comments.GhprbCommentFile;
 import org.jenkinsci.plugins.ghprb.extensions.status.GhprbSimpleStatus;
 import org.kohsuke.github.GHCommitState;
 
-import antlr.ANTLRException;
-import hudson.model.AbstractProject;
-import hudson.triggers.Trigger;
-import hudson.util.DescribableList;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
 
+public abstract class GhprbTriggerBackwardsCompatible extends Trigger<Job<?, ?>> {
 
-public abstract class GhprbTriggerBackwardsCompatible extends Trigger<AbstractProject<?, ?>> {
-    
     public abstract DescribableList<GhprbExtension, GhprbExtensionDescriptor> getExtensions();
-    
-    protected final int latestVersion = 3;
-    
+
+    protected final static int LATEST_VERSION = 3;
 
     protected Integer configVersion;
 
     public GhprbTriggerBackwardsCompatible(String cron) throws ANTLRException {
         super(cron);
     }
-    
+
 
     @Deprecated
     protected transient String commentFilePath;
+
     @Deprecated
     protected transient String msgSuccess;
+
     @Deprecated
     protected transient String msgFailure;
-    @Deprecated 
+
+    @Deprecated
     protected transient String commitStatusContext;
+
     @Deprecated
     protected transient GhprbGitHubAuth gitHubApiAuth;
+
     @Deprecated
     protected transient String project;
+
     @Deprecated
-    protected transient AbstractProject<?, ?> _project;
+    protected transient Job<?, ?> _project;
+
     @Deprecated
     protected transient Map<Integer, GhprbPullRequest> pullRequests;
-    
 
-    
-    
+
     protected void convertPropertiesToExtensions() {
         if (configVersion == null) {
             configVersion = 0;
         }
-        
+
         checkCommentsFile();
         checkBuildStatusMessages();
         checkCommitStatusContext();
-        
-        configVersion = latestVersion;
+
+        configVersion = LATEST_VERSION;
     }
-    
+
     private void checkBuildStatusMessages() {
         if (!StringUtils.isEmpty(msgFailure) || !StringUtils.isEmpty(msgSuccess)) {
             List<GhprbBuildResultMessage> messages = new ArrayList<GhprbBuildResultMessage>(2);
@@ -87,19 +90,19 @@ public abstract class GhprbTriggerBackwardsCompatible extends Trigger<AbstractPr
             commentFilePath = null;
         }
     }
-    
+
     private void checkCommitStatusContext() {
         if (configVersion < 1) {
             GhprbSimpleStatus status = new GhprbSimpleStatus(commitStatusContext);
             addIfMissing(status);
         }
     }
-    
+
     protected void addIfMissing(GhprbExtension ext) {
         if (getExtensions().get(ext.getClass()) == null) {
             getExtensions().add(ext);
         }
     }
 
-    
+
 }
