@@ -1,12 +1,11 @@
 package org.jenkinsci.plugins.ghprb.extensions;
 
+import hudson.model.Action;
+import hudson.model.Job;
 import org.jenkinsci.plugins.ghprb.GhprbCause;
 
-import hudson.model.AbstractProject;
-import hudson.model.Action;
-
 public interface GhprbBuildStep extends Action {
-    public static String buildStep = "GhprbBuildStep";
-    
-    public void onScheduleBuild(AbstractProject<?, ?> project, GhprbCause cause);
+    String BUILD_STEP = "GhprbBuildStep";
+
+    void onScheduleBuild(Job<?, ?> project, GhprbCause cause);
 }
