@@ -1,10 +1,8 @@
 package org.jenkinsci.plugins.ghprb.extensions;
 
-import hudson.model.AbstractBuild;
+import hudson.model.Run;
 import hudson.model.TaskListener;
 
 public interface GhprbCommentAppender {
-
-    public String postBuildComment(AbstractBuild<?, ?> build, TaskListener listener);
-    
+    String postBuildComment(Run<?, ?> build, TaskListener listener);
 }
