@@ -1,12 +1,8 @@
 package org.jenkinsci.plugins.ghprb.extensions.comments;
 
-import java.io.IOException;
-import java.util.List;
-
 import hudson.Extension;
+import hudson.model.Run;
 import hudson.model.TaskListener;
-import hudson.model.AbstractBuild;
-
 import org.jenkinsci.plugins.ghprb.Ghprb;
 import org.jenkinsci.plugins.ghprb.extensions.GhprbCommentAppender;
 import org.jenkinsci.plugins.ghprb.extensions.GhprbExtension;
@@ -16,29 +12,32 @@ import org.jenkinsci.plugins.ghprb.extensions.GhprbProjectExtension;
 import org.kohsuke.github.GHCommitState;
 import org.kohsuke.stapler.DataBoundConstructor;
 
+import java.io.IOException;
+import java.util.List;
+
 public class GhprbBuildLog extends GhprbExtension implements GhprbCommentAppender, GhprbProjectExtension, GhprbGlobalExtension {
 
     @Extension
     public static final DescriptorImpl DESCRIPTOR = new DescriptorImpl();
-    
+
     private final Integer logExcerptLines;
 
     @DataBoundConstructor
     public GhprbBuildLog(Integer logExcerptLines) {
         this.logExcerptLines = logExcerptLines;
     }
-    
+
     public Integer getLogExcerptLines() {
-        return logExcerptLines == null ? 0 : logExcerptLines;
+        return logExcerptLines == null ? Integer.valueOf(0) : logExcerptLines;
     }
 
-    public String postBuildComment(AbstractBuild<?, ?> build, TaskListener listener) {
-        
+    public String postBuildComment(Run<?, ?> build, TaskListener listener) {
+
         StringBuilder msg = new StringBuilder();
         GHCommitState state = Ghprb.getState(build);
 
         int numLines = getDescriptor().getLogExcerptLinesDefault(this);
-        
+
         if (state != GHCommitState.SUCCESS && numLines > 0) {
             // on failure, append an excerpt of the build log
             try {
@@ -66,7 +65,7 @@ public class GhprbBuildLog extends GhprbExtension implements GhprbCommentAppende
     public DescriptorImpl getDescriptor() {
         return DESCRIPTOR;
     }
-    
+
 
     public static final class DescriptorImpl extends GhprbExtensionDescriptor implements GhprbGlobalExtension {
 
@@ -74,7 +73,7 @@ public class GhprbBuildLog extends GhprbExtension implements GhprbCommentAppende
         public String getDisplayName() {
             return "Append portion of build log";
         }
-        
+
         public Integer getLogExcerptLinesDefault(GhprbBuildLog local) {
             Integer lines = Ghprb.getDefaultValue(local, GhprbBuildLog.class, "getLogExcerptLines");
             if (lines == null) {
