@@ -1,12 +1,11 @@
 package org.jenkinsci.plugins.ghprb.extensions.comments;
 
 import hudson.Extension;
-import hudson.model.AbstractBuild;
 import hudson.model.AbstractDescribableImpl;
-import hudson.model.TaskListener;
 import hudson.model.Descriptor;
+import hudson.model.Run;
+import hudson.model.TaskListener;
 import hudson.util.ListBoxModel;
-
 import org.apache.commons.lang.StringUtils;
 import org.jenkinsci.plugins.ghprb.Ghprb;
 import org.jenkinsci.plugins.ghprb.extensions.GhprbCommentAppender;
@@ -18,8 +17,9 @@ public class GhprbBuildResultMessage extends AbstractDescribableImpl<GhprbBuildR
 
     @Extension
     public static final DescriptorImpl DESCRIPTOR = new DescriptorImpl();
-    
+
     private final String message;
+
     private final GHCommitState result;
 
     @DataBoundConstructor
@@ -36,7 +36,7 @@ public class GhprbBuildResultMessage extends AbstractDescribableImpl<GhprbBuildR
         return result;
     }
 
-    public String postBuildComment(AbstractBuild<?, ?> build, TaskListener listener) {
+    public String postBuildComment(Run<?, ?> build, TaskListener listener) {
         StringBuilder msg = new StringBuilder();
 
         GHCommitState state = Ghprb.getState(build);
@@ -66,9 +66,8 @@ public class GhprbBuildResultMessage extends AbstractDescribableImpl<GhprbBuildR
     public DescriptorImpl getDescriptor() {
         return DESCRIPTOR;
     }
-    
-    public static class DescriptorImpl extends Descriptor<GhprbBuildResultMessage> {
 
+    public static class DescriptorImpl extends Descriptor<GhprbBuildResultMessage> {
         public boolean isApplicable(Class<?> type) {
             return true;
         }
@@ -80,11 +79,11 @@ public class GhprbBuildResultMessage extends AbstractDescribableImpl<GhprbBuildR
 
         public ListBoxModel doFillResultItems(@QueryParameter String result) {
             ListBoxModel items = new ListBoxModel();
-            GHCommitState[] results = new GHCommitState[] { GHCommitState.SUCCESS, GHCommitState.ERROR, GHCommitState.FAILURE };
+            GHCommitState[] results = new GHCommitState[] {GHCommitState.SUCCESS, GHCommitState.ERROR, GHCommitState.FAILURE};
             for (GHCommitState nextResult : results) {
 
                 items.add(nextResult.toString(), nextResult.toString());
-                if (result.toString().equals(nextResult)) {
+                if (result.equals(nextResult.toString())) {
                     items.get(items.size() - 1).selected = true;
                 }
             }
