@@ -2,9 +2,8 @@ package org.jenkinsci.plugins.ghprb.extensions.comments;
 
 import hudson.Extension;
 import hudson.model.Result;
+import hudson.model.Run;
 import hudson.model.TaskListener;
-import hudson.model.AbstractBuild;
-
 import org.jenkinsci.plugins.ghprb.Ghprb;
 import org.jenkinsci.plugins.ghprb.GhprbTrigger;
 import org.jenkinsci.plugins.ghprb.extensions.GhprbCommentAppender;
@@ -20,33 +19,28 @@ public class GhprbPublishJenkinsUrl extends GhprbExtension implements GhprbComme
 
     @Extension
     public static final DescriptorImpl DESCRIPTOR = new DescriptorImpl();
-    
+
     private final String publishedURL;
-    
+
     @DataBoundConstructor
     public GhprbPublishJenkinsUrl(String publishedURL) {
         this.publishedURL = publishedURL;
     }
-    
+
     public String getPublishedURL() {
         return publishedURL;
     }
 
-    public String postBuildComment(AbstractBuild<?, ?> build, TaskListener listener) {
-        StringBuilder msg = new StringBuilder();
-
-        msg.append("\nRefer to this link for build results (access rights to CI server needed): \n");
-        msg.append(generateCustomizedMessage(build));
-        msg.append("\n");
-        
-        return msg.toString();
+    public String postBuildComment(Run<?, ?> build, TaskListener listener) {
+        return "\nRefer to this link for build results (access rights to CI server needed): \n"
+                + generateCustomizedMessage(build) + "\n";
     }
 
     public boolean addIfMissing() {
         return false;
     }
 
-    private String generateCustomizedMessage(AbstractBuild<?, ?> build) {
+    private String generateCustomizedMessage(Run<?, ?> build) {
         GhprbTrigger trigger = Ghprb.extractTrigger(build);
         if (trigger == null) {
             return "";
@@ -73,7 +67,6 @@ public class GhprbPublishJenkinsUrl extends GhprbExtension implements GhprbComme
     }
 
     public static final class DescriptorImpl extends GhprbExtensionDescriptor implements GhprbGlobalExtension {
-
         @Override
         public String getDisplayName() {
             return "Add link to Jenkins";
@@ -82,6 +75,5 @@ public class GhprbPublishJenkinsUrl extends GhprbExtension implements GhprbComme
         public boolean addIfMissing() {
             return false;
         }
-        
     }
 }
