@@ -3,17 +3,10 @@ package org.jenkinsci.plugins.ghprb.extensions.status;
 import hudson.Extension;
 import hudson.Util;
 import hudson.matrix.MatrixProject;
+import hudson.model.Job;
+import hudson.model.Run;
 import hudson.model.TaskListener;
-import hudson.model.AbstractBuild;
-import hudson.model.AbstractProject;
-
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-
 import jenkins.model.Jenkins;
-
 import org.apache.commons.lang.StringUtils;
 import org.jenkinsci.plugins.ghprb.Ghprb;
 import org.jenkinsci.plugins.ghprb.GhprbCause;
@@ -33,18 +26,29 @@ import org.kohsuke.github.GHCommitState;
 import org.kohsuke.github.GHRepository;
 import org.kohsuke.stapler.DataBoundConstructor;
 
-public class GhprbSimpleStatus extends GhprbExtension implements GhprbCommitStatus, GhprbGlobalExtension,
-                               GhprbProjectExtension, GhprbGlobalDefault {
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
+public class GhprbSimpleStatus extends GhprbExtension implements
+        GhprbCommitStatus, GhprbGlobalExtension, GhprbProjectExtension, GhprbGlobalDefault {
 
     @Extension
-    public static final DescriptorImpl DESCRIPTOR = new DescriptorImpl();
+    public static final DescriptorImpl /*GhprbSimpleStatusDescriptor*/ DESCRIPTOR = new DescriptorImpl();
 
-    private final String commitStatusContext;
+    private String commitStatusContext;
+
     private final Boolean showMatrixStatus;
+
     private final String triggeredStatus;
+
     private final String startedStatus;
+
     private final String statusUrl;
+
     private final Boolean addTestResults;
+
     private final List<GhprbBuildResultMessage> completedStatus;
 
     public GhprbSimpleStatus() {
@@ -77,7 +81,7 @@ public class GhprbSimpleStatus extends GhprbExtension implements GhprbCommitStat
     }
 
     public boolean getShowMatrixStatus() {
-        return showMatrixStatus == null ? false: showMatrixStatus;
+        return showMatrixStatus == null ? false : showMatrixStatus;
     }
 
     public String getCommitStatusContext() {
@@ -93,7 +97,7 @@ public class GhprbSimpleStatus extends GhprbExtension implements GhprbCommitStat
     }
 
     public Boolean getAddTestResults() {
-        return addTestResults == null ? false : addTestResults;
+        return addTestResults == null ? Boolean.valueOf(false) : addTestResults;
     }
 
     public List<GhprbBuildResultMessage> getCompletedStatus() {
@@ -104,7 +108,8 @@ public class GhprbSimpleStatus extends GhprbExtension implements GhprbCommitStat
         return true;
     }
 
-    public void onBuildTriggered(AbstractProject<?, ?> project,
+    @Override
+    public void onBuildTriggered(Job<?, ?> project,
                                  String commitSha,
                                  boolean isMergeable,
                                  int prId,
@@ -119,7 +124,9 @@ public class GhprbSimpleStatus extends GhprbExtension implements GhprbCommitStat
         }
 
         String statusUrl = getDescriptor().getStatusUrlDefault(this);
-        String commitStatusContext = getDescriptor().getCommitStatusContextDefault(this);
+        if (commitStatusContext == "") {
+            commitStatusContext = getDescriptor().getCommitStatusContextDefault(this);
+        }
 
         String context = Util.fixEmpty(commitStatusContext);
         context = Ghprb.replaceMacros(project, context);
@@ -127,11 +134,11 @@ public class GhprbSimpleStatus extends GhprbExtension implements GhprbCommitStat
         if (!StringUtils.isEmpty(triggeredStatus)) {
             sb.append(Ghprb.replaceMacros(project, triggeredStatus));
         } else {
-            sb.append("Build triggered.");
+            sb.append("Build triggered");
             if (isMergeable) {
-                sb.append(" sha1 is merged.");
+                sb.append(" for merge commit.");
             } else {
-                sb.append(" sha1 is original commit.");
+                sb.append(" for original commit.");
             }
         }
 
@@ -148,14 +155,16 @@ public class GhprbSimpleStatus extends GhprbExtension implements GhprbCommitStat
         }
     }
 
-    public void onEnvironmentSetup(AbstractBuild<?, ?> build,
+    @Override
+    public void onEnvironmentSetup(Run<?, ?> build,
                                    TaskListener listener,
                                    GHRepository repo) throws GhprbCommitStatusException {
         // no need to create a commit here -- the onBuildStart() event will fire
         // soon and will respect's the user's settings for startedStatus.
     }
 
-    public void onBuildStart(AbstractBuild<?, ?> build,
+    @Override
+    public void onBuildStart(Run<?, ?> build,
                              TaskListener listener,
                              GHRepository repo) throws GhprbCommitStatusException {
         String startedStatus = getDescriptor().getStartedStatusDefault(this);
@@ -163,7 +172,7 @@ public class GhprbSimpleStatus extends GhprbExtension implements GhprbCommitStat
         // If the showMatrixStatus checkbox is selected and the job is not a Matrix Job (Children
         // nodes), then don't post statuses
         boolean showMatrixStatus = getDescriptor().getShowMatrixStatusDefault(this);
-        if(showMatrixStatus && !(build.getProject() instanceof MatrixProject)){
+        if (showMatrixStatus && !(build.getParent() instanceof MatrixProject)) {
             return;
         }
 
@@ -177,7 +186,7 @@ public class GhprbSimpleStatus extends GhprbExtension implements GhprbCommitStat
         if (StringUtils.isEmpty(startedStatus)) {
             sb.append("Build started");
             if (c != null) {
-                sb.append(c.isMerged() ? " sha1 is merged." : " sha1 is original commit.");
+                sb.append(c.isMerged() ? " for merge commit." : " for original commit.");
             }
         } else {
             sb.append(Ghprb.replaceMacros(build, listener, startedStatus));
@@ -186,7 +195,7 @@ public class GhprbSimpleStatus extends GhprbExtension implements GhprbCommitStat
         createCommitStatus(build, listener, sb.toString(), repo, GHCommitState.PENDING);
     }
 
-    public void onBuildComplete(AbstractBuild<?, ?> build,
+    public void onBuildComplete(Run<?, ?> build,
                                 TaskListener listener,
                                 GHRepository repo) throws GhprbCommitStatusException {
         List<GhprbBuildResultMessage> completedStatus = getDescriptor().getCompletedStatusDefault(this);
@@ -194,7 +203,7 @@ public class GhprbSimpleStatus extends GhprbExtension implements GhprbCommitStat
 
         // If the showMatrixStatus checkbox is selected and the job is not a Matrix Job (Children
         // nodes), then don't post statuses
-        if(showMatrixStatus && !(build.getProject() instanceof MatrixProject)) {
+        if (showMatrixStatus && !(build.getParent() instanceof MatrixProject)) {
             return;
         }
 
@@ -219,8 +228,8 @@ public class GhprbSimpleStatus extends GhprbExtension implements GhprbCommitStat
             listener.getLogger().println("Unable to get pull request builder trigger!!");
         } else {
             JobConfiguration jobConfiguration =
-                            JobConfiguration.builder()
-                                            .printStackTrace(trigger.getDisplayBuildErrorsOnDownstreamBuilds()).build();
+                    JobConfiguration.builder()
+                            .printStackTrace(trigger.getDisplayBuildErrorsOnDownstreamBuilds()).build();
 
             GhprbBuildManager buildManager = GhprbBuildManagerFactoryUtil.getBuildManager(build, jobConfiguration);
             if (getAddTestResults()) {
@@ -232,7 +241,7 @@ public class GhprbSimpleStatus extends GhprbExtension implements GhprbCommitStat
         createCommitStatus(build, listener, sb.toString(), repo, state);
     }
 
-    private void createCommitStatus(AbstractBuild<?, ?> build,
+    private void createCommitStatus(Run<?, ?> build,
                                     TaskListener listener,
                                     String message,
                                     GHRepository repo,
@@ -261,12 +270,13 @@ public class GhprbSimpleStatus extends GhprbExtension implements GhprbCommitStat
         context = Ghprb.replaceMacros(build, listener, context);
 
         listener.getLogger().println(String.format("Setting status of %s to %s with url %s and message: '%s'",
-                                                   sha1,
-                                                   state,
-                                                   url,
-                                                   message));
+                sha1,
+                state,
+                url,
+                message
+        ));
         if (context != null) {
-            listener.getLogger().println(String.format("Using context: " + context));
+            listener.getLogger().println("Using context: " + context);
         }
 
         try {
@@ -277,12 +287,12 @@ public class GhprbSimpleStatus extends GhprbExtension implements GhprbCommitStat
     }
 
     @Override
-    public DescriptorImpl getDescriptor() {
+    public DescriptorImpl /*GhprbSimpleStatusDescriptor*/ getDescriptor() {
         return DESCRIPTOR;
     }
 
-    public static final class DescriptorImpl extends GhprbExtensionDescriptor
-                                             implements GhprbGlobalExtension, GhprbProjectExtension {
+    public static final class DescriptorImpl extends GhprbExtensionDescriptor /*GhprbSimpleStatusDescriptor*/
+            implements GhprbGlobalExtension, GhprbProjectExtension {
 
         @Override
         public String getDisplayName() {
@@ -312,14 +322,13 @@ public class GhprbSimpleStatus extends GhprbExtension implements GhprbCommitStat
         public String getCommitStatusContextDefault(GhprbSimpleStatus local) {
             return Ghprb.getDefaultValue(local, GhprbSimpleStatus.class, "getCommitStatusContext");
         }
-        
-        public Boolean getShowMatrixStatusDefault(GhprbSimpleStatus local){
-            return Ghprb.getDefaultValue(local,GhprbSimpleStatus.class, "getShowMatrixStatus");
+
+        public Boolean getShowMatrixStatusDefault(GhprbSimpleStatus local) {
+            return Ghprb.getDefaultValue(local, GhprbSimpleStatus.class, "getShowMatrixStatus");
         }
 
         public boolean addIfMissing() {
             return false;
         }
-
     }
 }
