@@ -9,21 +9,45 @@ import java.util.List;
 
 class GhprbTriggerContext implements Context {
     List<String> admins = new ArrayList<String>();
+
     List<String> userWhitelist = new ArrayList<String>();
+
     List<String> orgWhitelist = new ArrayList<String>();
+
     List<GhprbBranch> whiteListTargetBranches = new ArrayList<GhprbBranch>();
+
     List<GhprbBranch> blackListTargetBranches = new ArrayList<GhprbBranch>();
+
+    List<String> blackListLabels = new ArrayList<String>();
+
+    List<String> whiteListLabels = new ArrayList<String>();
+
     String cron = "H/5 * * * *";
+
     String triggerPhrase;
+
     String skipBuildPhrase;
+
+    String blackListCommitAuthor;
+
     boolean onlyTriggerPhrase;
+
     boolean useGitHubHooks;
+
     boolean permitAll;
+
     boolean autoCloseFailedPullRequests;
+
     boolean allowMembersOfWhitelistedOrgsAsAdmin;
+
     boolean displayBuildErrorsOnDownstreamBuilds;
     boolean checkMergeCommit;
     String buildDescriptionTemplate;
+
+    String includedRegions;
+
+    String excludedRegions;
+
     GhprbExtensionContext extensionContext = new GhprbExtensionContext();
 
     /**
@@ -107,6 +131,40 @@ class GhprbTriggerContext implements Context {
     }
 
     /**
+     * Set label lists which are considered whitelisted for this specific job
+     */
+    private void whiteListLabel(String whiteListLabel) {
+        whiteListLabels.add(whiteListLabel);
+    }
+
+    /**
+     * Set label lists which are considered whitelisted for this specific job
+     */
+    public void whiteListLabels(Iterable<String> labels) {
+        for (String label : labels) {
+            whiteListLabel(label);
+        }
+
+    }
+
+    /**
+     * Set label lists which are considered blacklisted for this specific job
+     */
+    private void blackListLabel(String label) {
+        blackListLabels.add(label);
+    }
+
+    /**
+     * Set label lists which are considered blacklisted for this specific job
+     */
+    public void blackListLabels(Iterable<String> labels) {
+        for (String label : labels) {
+            blackListLabel(label);
+        }
+
+    }
+
+    /**
      * This schedules polling to GitHub for new changes in pull requests.
      */
     public void cron(String cron) {
@@ -128,11 +186,19 @@ class GhprbTriggerContext implements Context {
     }
 
     /**
+     * When filled, pull requests comits from this user will be skipped.
+     */
+    public void blackListCommitAuthor(String blackListCommitAuthor) {
+        this.blackListCommitAuthor = blackListCommitAuthor;
+    }
+
+    /**
      * When set, only commenting the trigger phrase in the pull request will trigger a build.
      */
     public void onlyTriggerPhrase(boolean onlyTriggerPhrase) {
         this.onlyTriggerPhrase = onlyTriggerPhrase;
     }
+
 
     /**
      * When set, only commenting the trigger phrase in the pull request will trigger a build.
@@ -229,7 +295,7 @@ class GhprbTriggerContext implements Context {
      * When filled, changes the default build description template
      */
     public void buildDescriptionTemplate(String template) {
-       this.buildDescriptionTemplate = template;
+        this.buildDescriptionTemplate = template;
     }
 
     /**
@@ -237,6 +303,30 @@ class GhprbTriggerContext implements Context {
      */
     public void extensions(Runnable closure) {
         ContextExtensionPoint.executeInContext(closure, extensionContext);
+    }
+
+    public void includedRegions(String regions) {
+        this.includedRegions = regions;
+    }
+
+    public void excludedRegions(String regions) {
+        this.excludedRegions = regions;
+    }
+
+    public void includedRegions(Iterable<String> regions) {
+        String includedRegionsStr = "";
+        for (String region : regions) {
+            includedRegionsStr += (region + "\n");
+        }
+        includedRegions(includedRegionsStr);
+    }
+
+    public void excludedRegions(Iterable<String> regions) {
+        String excludedRegionsStr = "";
+        for (String region : regions) {
+            excludedRegionsStr += (region + "\n");
+        }
+        excludedRegions(excludedRegionsStr);
     }
 
 }
