@@ -1,13 +1,13 @@
 package org.jenkinsci.plugins.ghprb.manager.impl;
 
-import hudson.model.AbstractBuild;
+import hudson.model.Run;
 
 /**
  * @author mdelapenya (Manuel de la Peña)
  */
 public class GhprbDefaultBuildManager extends GhprbBaseBuildManager {
 
-    public GhprbDefaultBuildManager(AbstractBuild<?, ?> build) {
+    public GhprbDefaultBuildManager(Run<?, ?> build) {
         super(build);
     }
 
