@@ -2,9 +2,10 @@ package org.jenkinsci.plugins.ghprb.upstream;
 
 import hudson.Extension;
 import hudson.Launcher;
-import hudson.model.*;
+import hudson.model.AbstractBuild;
+import hudson.model.BuildListener;
+import hudson.model.Descriptor;
 import hudson.tasks.BuildWrapper;
-
 import org.jenkinsci.plugins.ghprb.extensions.comments.GhprbBuildResultMessage;
 import org.kohsuke.github.GHCommitState;
 import org.kohsuke.stapler.DataBoundConstructor;
@@ -19,29 +20,37 @@ import java.util.Map.Entry;
 /**
  * @author Kevin Suwala
  */
-
 public class GhprbUpstreamStatus extends BuildWrapper {
+
+    private static final int INITIAL_CAPACITY = 5;
+
     private final Boolean showMatrixStatus;
+
     private final String commitStatusContext;
+
     private final String triggeredStatus;
+
     private final String startedStatus;
+
     private final String statusUrl;
+
     private final Boolean addTestResults;
+
     private final List<GhprbBuildResultMessage> completedStatus;
-    
+
     // sets the context and message as env vars so that they are available in the Listener class
     @Override
-    public void makeBuildVariables(@SuppressWarnings("rawtypes") AbstractBuild build, Map<String,String> variables){
-        variables.put("ghprbShowMatrixStatus",Boolean.toString(getShowMatrixStatus()));
+    public void makeBuildVariables(@SuppressWarnings("rawtypes") AbstractBuild build, Map<String, String> variables) {
+        variables.put("ghprbShowMatrixStatus", Boolean.toString(getShowMatrixStatus()));
         variables.put("ghprbUpstreamStatus", "true");
         variables.put("ghprbCommitStatusContext", getCommitStatusContext());
         variables.put("ghprbTriggeredStatus", getTriggeredStatus());
         variables.put("ghprbStartedStatus", getStartedStatus());
         variables.put("ghprbStatusUrl", getStatusUrl());
         variables.put("ghprbAddTestResults", Boolean.toString(getAddTestResults()));
-        
-        Map<GHCommitState, StringBuilder> statusMessages = new HashMap<GHCommitState, StringBuilder>(5);
-        
+
+        Map<GHCommitState, StringBuilder> statusMessages = new HashMap<GHCommitState, StringBuilder>(INITIAL_CAPACITY);
+
         for (GhprbBuildResultMessage message : getCompletedStatus()) {
             GHCommitState state = message.getResult();
             StringBuilder sb;
@@ -54,7 +63,7 @@ public class GhprbUpstreamStatus extends BuildWrapper {
             }
             sb.append(message.getMessage());
         }
-        
+
         for (Entry<GHCommitState, StringBuilder> next : statusMessages.entrySet()) {
             String key = String.format("ghprb%sMessage", next.getKey().name());
             variables.put(key, next.getValue().toString());
@@ -63,27 +72,30 @@ public class GhprbUpstreamStatus extends BuildWrapper {
 
     @Override
     @SuppressWarnings("unchecked")
-    public BuildWrapper.Environment setUp(@SuppressWarnings("rawtypes") AbstractBuild build, Launcher launcher, BuildListener listener) throws IOException, InterruptedException {
+    public BuildWrapper.Environment setUp(@SuppressWarnings("rawtypes") AbstractBuild build, Launcher launcher, BuildListener listener)
+            throws IOException, InterruptedException {
         makeBuildVariables(build, build.getBuildVariables());
-        return new Environment(){};
+        return new Environment() {
+        };
     }
 
     @Override
     @SuppressWarnings("unchecked")
-    public void preCheckout(@SuppressWarnings("rawtypes") AbstractBuild build, Launcher launcher, BuildListener listener) throws IOException, InterruptedException {
+    public void preCheckout(@SuppressWarnings("rawtypes") AbstractBuild build, Launcher launcher, BuildListener listener)
+            throws IOException, InterruptedException {
         makeBuildVariables(build, build.getBuildVariables());
     }
 
     @DataBoundConstructor
     public GhprbUpstreamStatus(
             Boolean showMatrixStatus,
-            String commitStatusContext, 
-            String statusUrl, 
-            String triggeredStatus, 
-            String startedStatus, 
+            String commitStatusContext,
+            String statusUrl,
+            String triggeredStatus,
+            String startedStatus,
             Boolean addTestResults,
             List<GhprbBuildResultMessage> completedStatus
-            ) {
+    ) {
         this.showMatrixStatus = showMatrixStatus;
         this.statusUrl = statusUrl;
         this.commitStatusContext = commitStatusContext == null ? "" : commitStatusContext;
@@ -92,12 +104,12 @@ public class GhprbUpstreamStatus extends BuildWrapper {
         this.addTestResults = addTestResults;
         this.completedStatus = completedStatus;
     }
-    
+
 
     public String getStatusUrl() {
         return statusUrl == null ? "" : statusUrl;
     }
-    
+
     public String getCommitStatusContext() {
         return commitStatusContext == null ? "" : commitStatusContext;
     }
@@ -105,17 +117,17 @@ public class GhprbUpstreamStatus extends BuildWrapper {
     public String getStartedStatus() {
         return startedStatus == null ? "" : startedStatus;
     }
-    
+
     public String getTriggeredStatus() {
         return triggeredStatus == null ? "" : triggeredStatus;
     }
 
     public Boolean getAddTestResults() {
-        return addTestResults == null ? false : addTestResults;
+        return addTestResults == null ? Boolean.valueOf(false) : addTestResults;
     }
 
-    public Boolean getShowMatrixStatus(){
-        return showMatrixStatus == null ? false : showMatrixStatus;
+    public Boolean getShowMatrixStatus() {
+        return showMatrixStatus == null ? Boolean.valueOf(false) : showMatrixStatus;
     }
 
 
@@ -131,6 +143,4 @@ public class GhprbUpstreamStatus extends BuildWrapper {
             return "Set GitHub commit status with custom context and message (Must configure upstream job using GHPRB trigger)";
         }
     }
-
-
 }
