@@ -1,9 +1,5 @@
 package org.jenkinsci.plugins.ghprb;
 
-import static org.fest.assertions.Assertions.assertThat;
-
-import java.util.List;
-
 import org.jenkinsci.plugins.ghprb.extensions.build.GhprbCancelBuildsOnUpdate;
 import org.jenkinsci.plugins.ghprb.extensions.comments.GhprbBuildLog;
 import org.jenkinsci.plugins.ghprb.extensions.comments.GhprbBuildResultMessage;
@@ -14,9 +10,13 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.runners.MockitoJUnitRunner;
 
+import java.util.List;
+
+import static org.fest.assertions.Assertions.assertThat;
+
 @RunWith(MockitoJUnitRunner.class)
 public class GeneralTest {
-    
+
     private void checkClassForGetters(Class<?> clazz) {
         List<String> errors = GhprbTestUtil.checkClassForGetters(clazz);
         assertThat(errors).isEmpty();
@@ -36,12 +36,11 @@ public class GeneralTest {
     public void checkPullRequestMergeForGetters() {
         checkClassForGetters(GhprbPullRequestMerge.class);
     }
-    
+
     @Test
     public void checkBuildLogForGetters() {
         checkClassForGetters(GhprbBuildLog.class);
     }
-
 
     @Test
     public void checkBuildResultMessageForGetters() {
@@ -58,7 +57,6 @@ public class GeneralTest {
         checkClassForGetters(GhprbCommentFile.class);
     }
 
-
     @Test
     public void checkSimpleStatusForGetters() {
         checkClassForGetters(GhprbSimpleStatus.class);
@@ -68,5 +66,4 @@ public class GeneralTest {
     public void checkCancelBuildsOnUpdateForGetters() {
         checkClassForGetters(GhprbCancelBuildsOnUpdate.class);
     }
-
 }
