@@ -1,21 +1,15 @@
 package org.jenkinsci.plugins.ghprb;
 
-import java.io.IOException;
-import java.io.PrintStream;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
-import java.lang.reflect.Field;
-
-import hudson.model.AbstractBuild;
-import hudson.model.FreeStyleBuild;
-import hudson.model.ItemGroup;
-import hudson.model.StreamBuildListener;
+import com.coravy.hudson.plugins.github.GithubProjectProperty;
+import hudson.FilePath;
+import hudson.Launcher;
 import hudson.model.FreeStyleProject;
-import hudson.model.Run;
+import hudson.model.ItemGroup;
 import hudson.model.Result;
-
+import hudson.model.Run;
+import hudson.model.TaskListener;
+import hudson.triggers.Trigger;
+import hudson.triggers.TriggerDescriptor;
 import org.jenkinsci.plugins.ghprb.GhprbTrigger.DescriptorImpl;
 import org.junit.After;
 import org.junit.Before;
@@ -35,16 +29,22 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.runners.MockitoJUnitRunner;
 
-import com.coravy.hudson.plugins.github.GithubProjectProperty;
+import java.io.File;
+import java.io.IOException;
+import java.io.PrintStream;
+import java.lang.reflect.Field;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
 
-import static org.fest.assertions.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.anyString;
 import static org.mockito.Matchers.anyInt;
+import static org.mockito.Matchers.anyString;
+import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
@@ -55,46 +55,63 @@ public class GhprbPullRequestMergeTest {
     public JenkinsRule jenkinsRule = new JenkinsRule();
 
     private FreeStyleProject project = mock(FreeStyleProject.class);
-    private AbstractBuild<?, ?> build = mock(FreeStyleBuild.class);
+
+    private Run<?, ?> build = mock(Run.class);
 
     @Mock
     private GhprbPullRequest pullRequest;
+
     @Mock
     private GHPullRequest pr;
 
     @Mock
     private GitUser committer;
+
     @Mock
     private GHUser triggerSender;
+
     @Mock
     private GHUser prCreator;
+
     @Mock
     private GhprbCause cause;
+
     @Mock
     private Ghprb helper;
+
     @Mock
     private GhprbRepository repo;
+
     @Mock
     private GHRepository ghRepo;
 
     @Mock
-    private StreamBuildListener listener;
+    private TaskListener listener;
 
     @Mock
     private ItemGroup<?> parent;
 
+    @Mock
+    private Launcher launcher;
+
+    private FilePath mockFilePath;
+
     private final String triggerPhrase = "ok to merge";
+
     private final String nonTriggerPhrase = "This phrase is not the trigger phrase";
 
     private final String adminList = "admin";
 
     private final String adminLogin = "admin";
+
     private final String nonAdminLogin = "nonadmin";
 
     private final String committerName = "committer";
+
     private final String nonCommitterName = "noncommitter";
-    
+
     private final String committerEmail = "committer@mail.com";
+
     private final String nonCommitterEmail = "noncommitter@mail.com";
 
     private final String mergeComment = "merge";
@@ -105,15 +122,19 @@ public class GhprbPullRequestMergeTest {
 
     @Before
     public void beforeTest() throws Exception {
+        launcher = mock(Launcher.class);
+
+        mockFilePath = new FilePath(new File(""));
+
         triggerValues = new HashMap<String, Object>(10);
         triggerValues.put("adminlist", adminList);
         triggerValues.put("triggerPhrase", triggerPhrase);
-        
-        GhprbTrigger trigger = GhprbTestUtil.getTrigger(triggerValues);
+
+        final GhprbTrigger trigger = GhprbTestUtil.getTrigger(triggerValues);
         Mockito.doReturn(repo).when(trigger).getRepository();
 
         ConcurrentMap<Integer, GhprbPullRequest> pulls = new ConcurrentHashMap<Integer, GhprbPullRequest>(1);
-        
+
         pulls.put(pullId, pullRequest);
         Map<String, ConcurrentMap<Integer, GhprbPullRequest>> jobs = new HashMap<String, ConcurrentMap<Integer, GhprbPullRequest>>(1);
         jobs.put("project", pulls);
@@ -123,8 +144,7 @@ public class GhprbPullRequestMergeTest {
         repo.addPullRequests(pulls);
         Mockito.doReturn(pullRequest).when(repo).getPullRequest(pullId);
         Mockito.doReturn(pr).when(repo).getActualPullRequest(pullId);
-        
-        
+
         GithubProjectProperty projectProperty = new GithubProjectProperty("https://github.com/jenkinsci/ghprb-plugin");
         DescriptorImpl descriptor = trigger.getDescriptor();
 
@@ -132,11 +152,14 @@ public class GhprbPullRequestMergeTest {
 
         given(parent.getFullName()).willReturn("");
 
+        Map<TriggerDescriptor, Trigger<?>> map = new HashMap<TriggerDescriptor, Trigger<?>>();
+        map.put(descriptor, trigger);
+
         given(project.getParent()).willReturn(parent);
-        given(project.getTrigger(GhprbTrigger.class)).willReturn(trigger);
+        given(project.getTriggers()).willReturn(map);
         given(project.getName()).willReturn("project");
         given(project.getProperty(GithubProjectProperty.class)).willReturn(projectProperty);
-        given(project.isDisabled()).willReturn(false);
+        given(project.isBuildable()).willReturn(true);
 
         given(build.getCause(GhprbCause.class)).willReturn(cause);
         given(build.getResult()).willReturn(Result.SUCCESS);
@@ -173,12 +196,18 @@ public class GhprbPullRequestMergeTest {
     }
 
     @SuppressWarnings("unchecked")
-    private void setupConditions(String prUserLogin, String triggerLogin, String committerName, String committerEmail, String comment) throws IOException {
+    private void setupConditions(
+            String prUserLogin,
+            String triggerLogin,
+            String committerName,
+            String committerEmail,
+            String comment
+    ) throws IOException {
         given(triggerSender.getLogin()).willReturn(triggerLogin);
         given(triggerSender.getName()).willReturn(committerName);
         given(triggerSender.getEmail()).willReturn(committerEmail);
         given(committer.getName()).willReturn(this.committerName);
-        
+
         given(prCreator.getLogin()).willReturn(prUserLogin);
         given(pr.getUser()).willReturn(prCreator);
 
@@ -200,22 +229,22 @@ public class GhprbPullRequestMergeTest {
 
         given(cause.getCommentBody()).willReturn(comment);
     }
-    
+
     private void setupConditions(String triggerLogin, String committerName, String committerEmail, String comment) throws IOException {
         setupConditions(nonCommitterName, triggerLogin, committerName, committerEmail, comment);
     }
-    
+
     private GhprbPullRequestMerge setupMerger(
-            boolean onlyAdminsMerge, 
+            boolean onlyAdminsMerge,
             boolean disallowOwnCode,
             boolean failOnNonMerge,
             boolean deleteOnMerge,
             boolean allowMergeWithoutTriggerPhrase
-            ) {
+    ) {
 
         GhprbPullRequestMerge merger = spy(new GhprbPullRequestMerge(
-                mergeComment, 
-                onlyAdminsMerge, 
+                mergeComment,
+                onlyAdminsMerge,
                 disallowOwnCode,
                 failOnNonMerge,
                 deleteOnMerge,
@@ -228,7 +257,7 @@ public class GhprbPullRequestMergeTest {
     }
 
     private GhprbPullRequestMerge setupMerger(
-            boolean onlyAdminsMerge, 
+            boolean onlyAdminsMerge,
             boolean disallowOwnCode) {
         return setupMerger(onlyAdminsMerge, disallowOwnCode, false, false, false);
     }
@@ -242,35 +271,35 @@ public class GhprbPullRequestMergeTest {
         GhprbPullRequestMerge merger = setupMerger(onlyAdminsMerge, disallowOwnCode);
 
         setupConditions(nonAdminLogin, committerName, committerEmail, triggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(true);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(1)).merge(mergeComment);
 
         setupConditions(adminLogin, nonCommitterName, nonCommitterEmail, triggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(true);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(2)).merge(mergeComment);
 
         setupConditions(adminLogin, committerName, committerEmail, nonTriggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(true);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(2)).merge(mergeComment);
 
         setupConditions(nonAdminLogin, nonCommitterName, nonCommitterEmail, triggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(true);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(3)).merge(mergeComment);
 
         setupConditions(nonAdminLogin, nonCommitterName, nonCommitterEmail, nonTriggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(true);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(3)).merge(mergeComment);
 
         setupConditions(adminLogin, nonCommitterName, nonCommitterEmail, nonTriggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(true);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(3)).merge(mergeComment);
 
         setupConditions(nonAdminLogin, nonCommitterName, nonCommitterEmail, nonTriggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(true);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(3)).merge(mergeComment);
 
         setupConditions(adminLogin, committerName, committerEmail, triggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(true);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(4)).merge(mergeComment);
     }
 
@@ -283,11 +312,11 @@ public class GhprbPullRequestMergeTest {
         GhprbPullRequestMerge merger = setupMerger(onlyAdminsMerge, disallowOwnCode);
 
         setupConditions(adminLogin, committerName, committerEmail, triggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(true);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(1)).merge(mergeComment);
 
         setupConditions(nonAdminLogin, committerName, committerEmail, triggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(false);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(1)).merge(mergeComment);
     }
 
@@ -300,11 +329,11 @@ public class GhprbPullRequestMergeTest {
         GhprbPullRequestMerge merger = setupMerger(onlyAdminsMerge, disallowOwnCode);
 
         setupConditions(adminLogin, committerName, committerEmail, triggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(true);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(1)).merge(mergeComment);
 
         setupConditions(adminLogin, committerName, committerEmail, nonTriggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(true);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(1)).merge(mergeComment);
     }
 
@@ -317,11 +346,11 @@ public class GhprbPullRequestMergeTest {
         GhprbPullRequestMerge merger = setupMerger(onlyAdminsMerge, disallowOwnCode);
 
         setupConditions(adminLogin, nonCommitterName, nonCommitterEmail, triggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(true);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(1)).merge(mergeComment);
 
         setupConditions(adminLogin, committerName, committerEmail, triggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(false);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(1)).merge(mergeComment);
     }
 
@@ -334,41 +363,56 @@ public class GhprbPullRequestMergeTest {
         GhprbPullRequestMerge merger = setupMerger(onlyAdminsMerge, disallowOwnCode);
 
         setupConditions(nonAdminLogin, nonAdminLogin, nonCommitterName, nonCommitterEmail, triggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(false);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(0)).merge(mergeComment);
 
         setupConditions(nonAdminLogin, adminLogin, committerName, committerEmail, triggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(false);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(0)).merge(mergeComment);
 
         setupConditions(nonAdminLogin, adminLogin, nonCommitterName, nonCommitterEmail, nonTriggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(true);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(0)).merge(mergeComment);
 
         setupConditions(nonAdminLogin, nonAdminLogin, nonCommitterName, nonCommitterEmail, triggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(false);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(0)).merge(mergeComment);
 
         setupConditions(nonAdminLogin, nonAdminLogin, nonCommitterName, nonCommitterEmail, nonTriggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(true);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(0)).merge(mergeComment);
 
         setupConditions(nonAdminLogin, adminLogin, committerName, committerEmail, nonTriggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(true);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(0)).merge(mergeComment);
 
         setupConditions(nonAdminLogin, nonAdminLogin, committerName, committerEmail, nonTriggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(true);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(0)).merge(mergeComment);
-        
+
         setupConditions(adminLogin, adminLogin, nonCommitterName, nonCommitterEmail, triggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(false);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(0)).merge(mergeComment);
 
         setupConditions(nonAdminLogin, adminLogin, nonCommitterName, nonCommitterEmail, triggerPhrase);
-        assertThat(merger.perform(build, null, listener)).isEqualTo(true);
+        merger.perform(build, mockFilePath, launcher, listener);
         verify(pr, times(1)).merge(mergeComment);
-        
     }
 
+    @Test
+    public void testFailOnNonMerge() throws Exception {
+
+        GhprbPullRequestMerge merger = setupMerger(
+            true,
+            true,
+            true,
+            false,
+            false
+        );
+
+        setupConditions(adminLogin, adminLogin, nonCommitterName, nonCommitterEmail, triggerPhrase);
+        merger.perform(build, mockFilePath, launcher, listener);
+
+        verify(build).setResult(Result.FAILURE);
+    }
 }
