@@ -1,8 +1,8 @@
 package org.jenkinsci.plugins.ghprb.manager.configuration;
 
-import static org.fest.assertions.Assertions.assertThat;
-
 import org.junit.Test;
+
+import static org.fest.assertions.Assertions.assertThat;
 
 /**
  * @author mdelapenya (Manuel de la Peña)
