@@ -1,21 +1,17 @@
 package org.jenkinsci.plugins.ghprb.manager.factory;
 
-import static org.fest.assertions.Assertions.assertThat;
-
 import com.cloudbees.plugins.flow.BuildFlow;
 import com.cloudbees.plugins.flow.FlowRun;
-
 import hudson.matrix.MatrixBuild;
 import hudson.matrix.MatrixProject;
-
 import org.jenkinsci.plugins.ghprb.manager.GhprbBuildManager;
-import org.jenkinsci.plugins.ghprb.manager.factory.GhprbBuildManagerFactoryUtil;
 import org.jenkinsci.plugins.ghprb.manager.impl.GhprbDefaultBuildManager;
 import org.jenkinsci.plugins.ghprb.manager.impl.downstreambuilds.BuildFlowBuildManager;
 import org.jenkinsci.plugins.ghprb.rules.JenkinsRuleWithBuildFlow;
-
 import org.junit.Rule;
 import org.junit.Test;
+
+import static org.fest.assertions.Assertions.assertThat;
 
 /**
  * @author mdelapenya (Manuel de la Peña)
@@ -46,5 +42,4 @@ public class GhprbBuildManagerFactoryUtilTest {
         // THEN
         assertThat(buildManager).isInstanceOf(BuildFlowBuildManager.class);
     }
-
 }
