@@ -1,15 +1,7 @@
 package org.jenkinsci.plugins.ghprb.manager.impl;
 
-import static org.fest.assertions.Assertions.assertThat;
-import static org.mockito.BDDMockito.given;
-
-import java.util.HashMap;
-import java.util.Map;
-
-
 import hudson.matrix.MatrixBuild;
 import hudson.matrix.MatrixProject;
-
 import org.jenkinsci.plugins.ghprb.GhprbITBaseTestCase;
 import org.jenkinsci.plugins.ghprb.GhprbTestUtil;
 import org.jenkinsci.plugins.ghprb.manager.GhprbBuildManager;
@@ -21,6 +13,12 @@ import org.junit.runner.RunWith;
 import org.jvnet.hudson.test.JenkinsRule;
 import org.mockito.runners.MockitoJUnitRunner;
 
+import java.util.HashMap;
+import java.util.Map;
+
+import static org.fest.assertions.Assertions.assertThat;
+import static org.mockito.BDDMockito.given;
+
 /**
  * @author mdelapenya (Manuel de la Peña)
  */
@@ -29,23 +27,23 @@ public class GhprbDefaultBuildManagerTest extends GhprbITBaseTestCase {
 
     @Rule
     public JenkinsRule jenkinsRule = new JenkinsRule();
-    
+
     private MatrixProject project;
 
     @Before
     public void setUp() throws Exception {
         // GhprbTestUtil.mockGithubUserPage();
         project = jenkinsRule.getInstance().createProject(MatrixProject.class, "MTXPRJ");
-       
-        Map<String, Object> config = new HashMap<String, Object>(1);
-       
+
+        Map<String, Object> config = new HashMap<>(1);
+
         config.put("publishedURL", "defaultPublishedURL");
-        super.beforeTest(config,  null, project);
+        super.beforeTest(config, null, project);
     }
 
     @Test
     public void shouldCalculateUrlFromDefault() throws Exception {
-        
+
         // GIVEN
         givenThatGhprbHasBeenTriggeredForAMatrixProject();
 
@@ -63,7 +61,7 @@ public class GhprbDefaultBuildManagerTest extends GhprbITBaseTestCase {
 
     private void givenThatGhprbHasBeenTriggeredForAMatrixProject() throws Exception {
         given(commitPointer.getSha()).willReturn("sha");
-        
+
         given(ghPullRequest.getNumber()).willReturn(1);
 
         given(ghRepository.getPullRequest(1)).willReturn(ghPullRequest);
